// graphql/src/generated/graphql.tsx
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
};

export type Affirmation = {
  __typename?: 'Affirmation';
  from: Scalars['ID'];
  to: Scalars['ID'];
};

export type Article = {
  __typename?: 'Article';
  body?: Maybe<Scalars['String']>;
  id?: Maybe<Scalars['ID']>;
  title?: Maybe<Scalars['String']>;
  user?: Maybe<User>;
};

export type ArticleInputArgs = {
  body: Scalars['String'];
  title: Scalars['String'];
  userId: Scalars['Int'];
};